5 ДЗ на проверку
