+++ js/facilities.js
@@ -1,0 +1,34 @@
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+];
+const NAMES = [
+  'Александр',
+  'Виктория',
+  'Дмитрий',
+  'Ульяна',
+  'Владимир',
+  'Яна',
+  'Григорий',
+  'Зоя',
+];
+const DESCIPTION = [
+  'Прекрасный день для прогулки!',
+  'Взгляни на эту красоту природы!',
+  'Очаровательный закат',
+  'Это место - настоящий рай на земле!',
+  'Какая атмосфера!',
+  'Подарок для глаз и души!',
+  'Как прекрасен этот мир!',
+  'Моменты счастья на каждом шагу.',
+  'Подарок от природы.',
+  'Это место наполнено энергией и покоем.',
+  'Как же здесь красиво!',
+  'Давайте насладимся этой красотой вместе!',
+  'Великолепная атмосфера для творчества.',
+  'Этот вид заставляет забыть обо всём на свете.',
+];
+
+export {MESSAGE, NAMES, DESCIPTION};
